perform some path sanitization over the generated names (#258)
* perform some path sanitization over the generated names

The current scheme for the file generation uses the function label,
which uses ':' as a parameter separator.  However, this character is not
valid on all file systems.  This results in partial content generation
on Windows.  Add a sanitization step which replaces "invalid" characters
with `_` instead.  This does not impact the rendering, only the
filenames and the generated URL references between the pages.  With
this, we can generate and link function documentation on Windows.

* Apply suggestion from code review

Co-authored-by: Mattt <mattt@github.com>