Update Package@swift-5.5.swift (#278)
* Update Package@swift-5.5.swift

Pin SwiftSemantics to main.  This is required since the 0.3.1 release of SwiftSemantics pins SwiftSyntax to 4ae758ab85ed2a5d3e3e8b5050a8ce52179bd102 and SwiftSyntaxHighlighter requires release/5.5.  In order to restore the ability to resolve the dependencies, we need to bump SwiftSemantics to main.

* Update Package@swift-5.5.swift

Use swift-syntax ~> 0.3.2

* Update Package@swift-5.5.swift

Use `revision` instead of `uptToNextMinor` for SwiftSemantics.  Bump SwiftSyntaxHighlighter.  This allows the package dependencies to be resolved on Windows.

Co-authored-by: Mattt <mattt@me.com>