build: add a GHA step for packaging swift-doc on Windows (#280)
This adds a GitHub Action step to package swift-doc for Windows for easy
distribution.  It installs to
`[WindowsVolume]\Library\Developer\SwiftDoc` by default.  This is
currently not configurable, though it likely should be.  However, this
is the simplest configuration possible and co-locates the swift-doc
package with the location of the Swift toolchain on Windows.

Update the publish CI job to build the MSI and upload the asset as a
release artifact.